+++ index.html
@@ -233,8 +233,8 @@
   </template>
 
   <!-- Подключение JavaScript-->
-<script src="/js/functions.js"></script>
-<script src="/js/main.js"></script>
+
+<script src="/js/main.js" type="module"></script>
 
 </body>
 </html>
